Fixed some awkward stats box bugs
Because updating the stats box involves setting some delays to have it try a few times, it was possible to get weird behaviour if someone jumped between items in the zoom to districts list too quickly.  Specifically, there could be duelling updates between the previous and current district.  This commit fixes that.