Simplified branching logic
Since the only difference between House and Senate districts is the value of two strings, I've reduced this to a single funtion definition, which just uses conditional logic to choose those strings.  This is not directly related to the bug, but I find the code easier to follow this way and it will simplify future maintenance.